Use dedicated classes for singleton and lru caches
When binding context through closures, each function needs to a store a pointer to the context. With classes, only one pointer to the prototype needs to be stored per class.

Per-object memory cost matters when creating lots of memoized objects. For our use cases with 100k+ objects, this reduces memory usage from ~800MB to ~650MB.